feat(core): declare module MidwayCore
for merge declaration of IMidwayBaseContext
Usage:
```ts
import { Context } from '@midwayjs/web'
declare module 'MidwayCore' {
  interface IMidwayBaseContext {
    foo: string
  }
}

declare const ctx: Context
ctx.foo
```
